+++ exinst-aeson/pkg.nix
@@ -1,17 +1,18 @@
-{ mkDerivation, aeson, base, bytestring, constraints, exinst, lib
-, QuickCheck, singletons, tasty, tasty-quickcheck
+{ mkDerivation, aeson, base, bytestring, constraints, exinst
+, exinst-base, lib, QuickCheck, singletons, tasty, tasty-quickcheck
 }:
 mkDerivation {
   pname = "exinst-aeson";
-  version = "0.7.1";
+  version = "0.9";
   src = ./.;
   libraryHaskellDepends = [
     aeson base constraints exinst singletons
   ];
   testHaskellDepends = [
-    aeson base bytestring exinst QuickCheck tasty tasty-quickcheck
+    aeson base bytestring exinst exinst-base QuickCheck tasty
+    tasty-quickcheck
   ];
   homepage = "https://github.com/k0001/exinst";
-  description = "Dependent pairs and their instances";
+  description = "@exinst@ support for @aeson@ package";
   license = lib.licenses.bsd3;
 }