+++ exinst-bytes/tests/Main.hs
@@ -25,6 +25,7 @@ import Test.Tasty.QuickCheck ((===))
 import qualified Test.Tasty.QuickCheck as QC
 
 import Exinst
+import Exinst.Base ()
 import Exinst.Bytes ()
 import Exinst.Cereal ()