+++ exinst-cereal/tests/Main.hs
@@ -22,6 +22,7 @@ import Test.Tasty.QuickCheck ((===))
 import qualified Test.Tasty.QuickCheck as QC
 
 import Exinst
+import Exinst.Base ()
 import Exinst.Cereal ()
 
 --------------------------------------------------------------------------------