+++ exinst-serialise/pkg.nix
@@ -1,15 +1,16 @@
-{ mkDerivation, base, binary, constraints, exinst, lib, QuickCheck
-, serialise, singletons, tasty, tasty-quickcheck
+{ mkDerivation, base, binary, constraints, exinst, exinst-base, lib
+, QuickCheck, serialise, singletons, tasty, tasty-quickcheck
 }:
 mkDerivation {
   pname = "exinst-serialise";
-  version = "0.7.1";
+  version = "0.9";
   src = ./.;
   libraryHaskellDepends = [
     base constraints exinst serialise singletons
   ];
   testHaskellDepends = [
-    base binary exinst QuickCheck serialise tasty tasty-quickcheck
+    base binary exinst exinst-base QuickCheck serialise tasty
+    tasty-quickcheck
   ];
   homepage = "https://github.com/k0001/exinst";
   description = "Dependent pairs and their instances";